removed blog app, moved profile model to dashboard
