--- blog/static/css/blog_detail.css
@@ -1,221 +1,0 @@
-/*
-* Prefixed by https://autoprefixer.github.io
-* PostCSS: v7.0.29,
-* Autoprefixer: v9.7.6
-* Browsers: last 15 version
-*/
-
-/*************************************** 
-             PAGE DEFAULTS
-****************************************/
-body {
-    background: whitesmoke;
-}
-
-code {
-    background-color: #E8E8E8 !important;
-    color: black !important;
-    padding: 2px 5px 2px 5px;
-    border-radius: 4px;
-}
-
-.blockquote {
-    border-left: var(--primary-color) solid 5px;
-    padding-left: 1rem;
-    position: relative;
-}
-
-/*************************************** 
-            END PAGE DEFAULTS
-****************************************/
-
-/*************************************** 
-            PROFILE CARD
-****************************************/
-.circle {
-    height: 100px;
-    width: 100px;
-    border-radius: 50%;
-    overflow: hidden;
-}
-
-.circle img {
-    height: 100%;
-    -o-object-fit: cover;
-       object-fit: cover;
-    -o-object-position: center;
-       object-position: center;
-}
-/*************************************** 
-            END PROFILE CARD
-****************************************/
-
-/*************************************** 
-            SIMILAR PROJECTS
-****************************************/
-.post-link {
-    display: inline-block;
-    background-color: whitesmoke;
-}
-
-.post-link:hover {
-    -webkit-box-shadow: 0 .125rem .25rem rgba(0,0,0,.075);
-            box-shadow: 0 .125rem .25rem rgba(0,0,0,.075);
-    text-decoration: none;
-}
-
-.read-more:hover,
-.text-body:hover {
-    text-decoration: none;
-}
-
-.read-more {
-    color: var(--primary-color, #f77171);
-}
-
-.read-more:hover {
-    color: var(--primary-color, #f77171);
-}
-
-#similar-projects .card-title {
-    font-weight: bold;
-}
-
-.main-content > p {
-    margin: 0;
-}
-
-.main-content pre {
-    margin: 0;
-}
-
-.main-title > h2,
-.main-title > h3,
-.main-title > h4,
-.main-title > h5,
-.main-title > h6 {
-    color: #000733;
-}
-/*************************************** 
-           END SIMILAR PROJECTS
-****************************************/
-.main-title {
-    color: #000733;
-}
-
-.ql-syntax {
-    padding: 1.75em;
-    background: #011627;
-    color: #d6deeb;
-    border-radius: 8px;
-}
-.card-text {
-    overflow-x: auto;
-}
-
-/*************************************** 
-            SOCIAL SHARE BUTTONS
-****************************************/
-.resp-sharing-button__link,
-.resp-sharing-button__icon {
-  display: inline-block
-}
-
-.resp-sharing-button__icon {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.resp-sharing-button__link {
-  text-decoration: none;
-  color: #fff;
-  margin: 0.5em
-}
-
-.resp-sharing-button__link:hover {
-  text-decoration: none;
-  color: #fff;
-}
-
-.resp-sharing-button {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: center;
-  border-radius: 5px;
-  transition: 25ms ease-out;
-  padding: 0.5em 0.75em;
-  font-family: Helvetica Neue,Helvetica,Arial,sans-serif
-}
-
-.resp-sharing-button__icon svg {
-  width: 1em;
-  height: 1em;
-  margin-right: 0.4em;
-  vertical-align: top
-}
-
-.resp-sharing-button--small svg {
-  margin: 0;
-  vertical-align: middle
-}
-
-/* Non solid icons get a stroke */
-.resp-sharing-button__icon {
-  stroke: #fff;
-  fill: none
-}
-
-/* Solid icons get a fill */
-.resp-sharing-button__icon--solid,
-.resp-sharing-button__icon--solidcircle {
-  fill: #fff;
-  stroke: none
-}
-
-.resp-sharing-button--twitter {
-  background-color: #55acee
-}
-
-.resp-sharing-button--twitter:hover {
-  background-color: #2795e9
-}
-
-.resp-sharing-button--facebook {
-  background-color: #3b5998
-}
-
-.resp-sharing-button--facebook:hover {
-  background-color: #2d4373
-}
-
-.resp-sharing-button--email {
-  background-color: #777
-}
-
-.resp-sharing-button--email:hover {
-  background-color: #5e5e5e
-}
-
-.resp-sharing-button--facebook {
-  background-color: #3b5998;
-  border-color: #3b5998;
-}
-
-.resp-sharing-button--facebook:hover,
-.resp-sharing-button--facebook:active {
-  background-color: #2d4373;
-  border-color: #2d4373;
-}
-
-.resp-sharing-button--twitter {
-  background-color: #55acee;
-  border-color: #55acee;
-}
-
-.resp-sharing-button--twitter:hover,
-.resp-sharing-button--twitter:active {
-  background-color: #2795e9;
-  border-color: #2795e9;
-}
-
