--- blog/static/css/blog_list.css
@@ -1,69 +1,0 @@
-/*
-* Prefixed by https://autoprefixer.github.io
-* PostCSS: v7.0.29,
-* Autoprefixer: v9.7.6
-* Browsers: last 15 version
-*/
-
-body {
-    background: whitesmoke;
-}
-
-/*************************************** 
-           BLOG LISTING CARDS
-****************************************/
-article {
-    background: white;
-    height: 100%;
-}
-
-.img-wrapper {
-    overflow: hidden;
-    height: 250px;
-}
-
-.img-wrapper img {
-    -o-object-fit: cover;
-       object-fit: cover;
-    -o-object-position: center;
-       object-position: center;
-    height: 100%;
-    width: 100%;
-}
-
-article .text-body {
-    height: 100%;
-}
-
-.blog-link:hover {
-    text-decoration: none;
-}
-
-.card-body,
-.text-body {
-    display: -webkit-box;
-    display: -webkit-flex;
-    display: -ms-flexbox;
-    display: flex;
-    -webkit-box-orient: vertical;
-    -webkit-box-direction: normal;
-    -webkit-flex-direction: column;
-        -ms-flex-direction: column;
-            flex-direction: column;
-}
-
-.card-text,
-.card-body {
-    -webkit-box-flex: 1;
-    -webkit-flex: 1;
-        -ms-flex: 1;
-            flex: 1;
-}
-
-hr {
-    width: 100%;
-}
-
-/*************************************** 
-           BLOG LISTING CARDS
-****************************************/
