--- blog/static/js/blog.js
@@ -1,2 +1,0 @@
-
-
